# backend/Dtos/UserDto.cs
namespace Backend.Dtos;

public sealed record UserResponse(
    Guid Id,
    string Username,
    string? Name,
    string? Address,
    string? PhoneNumber,
    DateTime CreatedAt
);

public sealed record UserUpdateRequest(
    string? Username,
    string? Name,
    string? Address,
    string? PhoneNumber
);